Add files via upload
This web app is an RGB guessing game. The user would start the game, and depending on the difficulty, 3 or 6 randomly colored squares will appear, with one of the squares having the same RGB property as displayed in the header. The objective is to guess which square has the displayed RGB property. The high score reflects how many consecutive correct guesses the user has achieved.